fix: separate out the api version base route function
